consolidating cli utils into index file - changed formatting for tree output
