+++ src/cli/utils/index.ts
@@ -1,8 +1,14 @@
 import { readdirSync } from 'node:fs';
 import path from 'node:path';
+import fs from 'node:fs';
+import os from 'node:os';
+
 import chalk from 'chalk';
 import { getReasonPhrase } from 'http-status-codes';
+import yargsParser from 'yargs-parser';
+
 import {
+  createConfig,
   ConfigValidationError,
   createLogger,
   formatBytes,
@@ -10,8 +16,8 @@ import {
   HttpError,
   type NetStorageFile,
   type WinstonLogLevel,
+  type NetStorageClientConfig,
 } from '@/index';
-import yargsParser from 'yargs-parser';
 
 /**
  * Specifies how each positional argument of a command should be resolved.
@@ -415,4 +421,155 @@ export function resolveCliArgs(
   }
 
   return result;
+}
+
+/**
+ * Loads project-level NetStorage configuration from `netstorage.json` in the current working directory.
+ *
+ * @returns {Promise<Partial<NetStorageClientConfig>>} Parsed config object or an empty object if not found.
+ */
+async function loadFileConfig(): Promise<Partial<NetStorageClientConfig>> {
+  const fallbackFiles = ['netstorage.json'];
+  for (const file of fallbackFiles) {
+    const fullPath = path.resolve(process.cwd(), file);
+    if (fs.existsSync(fullPath)) {
+      const raw = fs.readFileSync(fullPath, 'utf-8');
+      return JSON.parse(raw);
+    }
+  }
+
+  return {};
+}
+
+/**
+ * Removes any properties from the given object that have an undefined value.
+ *
+ * @example
+ * const input = { foo: 'bar', baz: undefined };
+ * const output = removeUndefined(input);
+ * // output is { foo: 'bar' }
+ *
+ * @param obj The object to filter.
+ * @returns A new object missing any properties with undefined values from the input object.
+ */
+function removeUndefined<T extends object>(obj: T): Partial<T> {
+  return Object.fromEntries(
+    Object.entries(obj).filter(([, v]) => v !== undefined),
+  ) as Partial<T>;
+}
+
+/**
+ * Resolves a complete NetStorage client config from layered configuration sources.
+ *
+ * Resolution order (highest to lowest priority):
+ * 1. CLI options (explicit flags passed to CLI command)
+ * 2. Environment variables (NETSTORAGE_* keys)
+ * 3. Project-level config file (`netstorage.json` in the current directory)
+ * 4. Persistent config (`~/.config/netstorage/.netstorage.json`)
+ *
+ * @param {Partial<NetStorageClientConfig>} [cliOptions] - Optional config overrides from CLI flags.
+ * @returns {Promise<NetStorageClientConfig>} Fully resolved configuration object.
+ */
+export async function loadClientConfig(
+  cliOptions: Partial<NetStorageClientConfig> = {},
+): Promise<NetStorageClientConfig> {
+  const envConfig: Partial<NetStorageClientConfig> = {
+    key: process.env.NETSTORAGE_API_KEY,
+    keyName: process.env.NETSTORAGE_API_KEYNAME,
+    host: process.env.NETSTORAGE_HOST,
+    ssl: process.env.NETSTORAGE_SSL === 'true',
+    cpCode: process.env.NETSTORAGE_CP_CODE,
+  };
+  const fileConfig = await loadFileConfig();
+  const persistentConfig = loadPersistentConfig();
+
+  // Merge: CLI > Env > File > Persistent
+  const merged: Partial<NetStorageClientConfig> = {
+    ...removeUndefined(persistentConfig),
+    ...removeUndefined(fileConfig),
+    ...removeUndefined(envConfig),
+    ...removeUndefined(cliOptions),
+  };
+
+  return createConfig(merged as NetStorageClientConfig);
+}
+
+const CONFIG_FILE = path.join(
+  os.homedir(),
+  '.config',
+  'netstorage',
+  'config.json',
+);
+
+/**
+ * Loads the persisted NetStorage client configuration from the user's config file.
+ *
+ * @returns {Partial<NetStorageClientConfig>} The saved configuration, or an empty object if none exists.
+ */
+export function loadPersistentConfig(): Partial<NetStorageClientConfig> {
+  if (fs.existsSync(CONFIG_FILE)) {
+    const raw = fs.readFileSync(CONFIG_FILE, 'utf-8');
+    return JSON.parse(raw);
+  }
+  return {};
+}
+
+/**
+ * Merges and saves the provided configuration with the existing persisted config.
+ * Creates the config directory if it does not exist.
+ *
+ * @param {Partial<NetStorageClientConfig>} update - Configuration values to persist.
+ */
+export function savePersistentConfig(
+  update: Partial<NetStorageClientConfig>,
+): void {
+  const current = loadPersistentConfig();
+  const merged = { ...current, ...update };
+  fs.mkdirSync(path.dirname(CONFIG_FILE), { recursive: true });
+  fs.writeFileSync(CONFIG_FILE, JSON.stringify(merged, null, 2), 'utf-8');
+}
+
+/**
+ * Overwrites the persistent configuration file with the given config object.
+ *
+ * @param {Partial<NetStorageClientConfig>} config - The complete configuration to persist.
+ */
+function overwritePersistentConfig(
+  config: Partial<NetStorageClientConfig>,
+): void {
+  fs.mkdirSync(path.dirname(CONFIG_FILE), { recursive: true });
+  fs.writeFileSync(CONFIG_FILE, JSON.stringify(config, null, 2), 'utf-8');
+}
+
+/**
+ * Deletes the persisted NetStorage configuration file if it exists.
+ */
+export function clearPersistentConfig(): void {
+  if (fs.existsSync(CONFIG_FILE)) {
+    fs.unlinkSync(CONFIG_FILE);
+  }
+}
+
+/**
+ * Deletes a specific key from the persisted NetStorage configuration file.
+ *
+ * @param {keyof NetStorageClientConfig} key - The configuration key to remove.
+ */
+export function clearPersistentConfigKey(
+  key: keyof NetStorageClientConfig,
+): void {
+  const current = loadPersistentConfig();
+  if (key in current) {
+    delete current[key];
+    overwritePersistentConfig(current);
+  }
+}
+
+/**
+ * Returns the path to the persistent configuration file.
+ *
+ * @returns {string} The absolute path to the saved config file.
+ */
+export function getPersistentConfigPath(): string {
+  return CONFIG_FILE;
 }
